24/3/4 ia-Feature: DB: New parameter delete_flag is introduced to control the deletion of current data in DB

+++ ia_selenium/ia_scrap.py
@@ -58,12 +58,12 @@ def ia_app(wd, confs, thread_name="Main", recursive=0):
             ia_app(wd, confs, thread_name)
 
 
-def save_csv_to_db(control_unit, files, tables, company):
+def save_csv_to_db(confs, files, tables, company):
     """
     This method saves data from CSV files into a database. It performs various operations based on
     the control_unit parameter to determine which tables to update or delete.
 
-    :param control_unit: an integer representing the control unit for determining which tables to update or delete
+    :param confs: dictionary with all confs
     :param files: a dictionary containing file paths for different CSV files.
     :param tables: a dictionary containing table names for different data types.
     :param company: a string representing the company name.
@@ -93,22 +93,26 @@ def save_csv_to_db(control_unit, files, tables, company):
     else:
         print("Database connection successful!")
         batch_size = 1000
+        control_unit = confs['control_unit']
         db_method.save_recover(cursor,
                                zip(tables['recover'], [None] * len(tables['recover']),
                                    [company] * len(tables['recover'])))
         db_method.save_data_into_db(cursor, files['contracts'], db_method.save_contract_history, batch_size)
-        db_method.delete_current_contract(cursor, company)
+        if confs['delete_flag']:
+            db_method.delete_current_contract(cursor, company)
 
         if control_unit & 1:
             # delete current table of fund & saving for later insertion
-            db_method.delete_current_fund_saving(cursor, company)
+            if confs['delete_flag']:
+                db_method.delete_current_fund_saving(cursor, company)
 
             # save saving & fund history
             db_method.save_data_into_db(cursor, files['saving'], db_method.save_saving_history, batch_size)
             db_method.save_data_into_db(cursor, files['fund'], db_method.save_fund_history, batch_size)
         if control_unit & 2:
             # delete current table of transaction for later insertion
-            db_method.delete_current_transaction(cursor, company)
+            if confs['delete_flag']:
+                db_method.delete_current_transaction(cursor, company)
 
             # save transaction history
             db_method.save_data_into_db(cursor, files['transaction'], db_method.save_transaction_history, batch_size)
@@ -117,8 +121,9 @@ def save_csv_to_db(control_unit, files, tables, company):
             # if there is no new contracts
             # otherwise just extend the table
             # if not new_contracts:
-            db_method.delete_current_participant_beneficiary(cursor, company)
-            db_method.delete_current_client(cursor, company)
+            if confs['delete_flag']:
+                db_method.delete_current_participant_beneficiary(cursor, company)
+                db_method.delete_current_client(cursor, company)
             db_method.save_data_into_db(cursor, files['client'], db_method.save_client_history, batch_size)
             db_method.save_data_into_db(cursor, files['participant'], db_method.save_participant_history, batch_size)
             db_method.save_data_into_db(cursor, files['beneficiary'], db_method.save_beneficiary_history, batch_size)
